requirements.txt
python-docx
bs4
requests
lxml
unidecode
gtts
mutagen
